Support testing with GitHub Codespace and local Docker
Use the "Run Locally" VS Code task to spin up a container that runs Jekyll to
generate and serve your site.  On your local computer, you'll need Docker
installed for this to work.  If you create a GitHub Codespace from this
repo (or an up-to-date fork), it will work out-of-the-box.